settings packge in common removed
