--- platform-sdk/swirlds-common/src/main/java/com/swirlds/common/settings/SettingsException.java
@@ -1,49 +1,0 @@
-/*
- * Copyright (C) 2016-2023 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.swirlds.common.settings;
-
-/**
- * This exception is thrown when there is a settings problem.
- *
- * @deprecated will be replaced by the {@link com.swirlds.config.api.Configuration} API in near future. If you need
- * 		to use this class please try to do as less static access as possible.
- */
-@Deprecated(forRemoval = true)
-public class SettingsException extends RuntimeException {
-
-    public SettingsException() {}
-
-    public SettingsException(final String message) {
-        super(message);
-    }
-
-    public SettingsException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-
-    public SettingsException(final Throwable cause) {
-        super(cause);
-    }
-
-    public SettingsException(
-            final String message,
-            final Throwable cause,
-            final boolean enableSuppression,
-            final boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
