+++ platform-sdk/swirlds-config-impl/build.gradle.kts
@@ -15,19 +15,20 @@
  */
 
 plugins {
-  id("com.swirlds.platform.conventions")
-  id("com.swirlds.platform.library")
-  id("com.swirlds.platform.maven-publish")
+    id("com.swirlds.platform.conventions")
+    id("com.swirlds.platform.library")
+    id("com.swirlds.platform.maven-publish")
 }
 
 dependencies {
-  // Individual Dependencies
-  implementation(project(":swirlds-base"))
-  api(project(":swirlds-common"))
-  api(project(":swirlds-config-api"))
-  api(libs.bundles.logging.api)
-  compileOnly(libs.spotbugs.annotations)
+    // Individual Dependencies
+    implementation(project(":swirlds-base"))
+    api(project(":swirlds-common"))
+    api(project(":swirlds-config-api"))
+    api(libs.bundles.logging.api)
+    compileOnly(libs.spotbugs.annotations)
 
-  // Test Dependencies
-  testImplementation(testLibs.bundles.junit)
+    // Test Dependencies
+    testImplementation(testLibs.bundles.junit)
+    testImplementation(project(":swirlds-unit-tests:common:swirlds-test-framework"))
 }
